feat(graphql): refactor user resolvers to use a typed Resolver interface for better type safety feat(prisma): create a singleton PrismaClient instance to optimize database connections in development feat(types): add Context and AuthData interfaces to improve type definitions for GraphQL resolvers and context chore(api): clean up unused imports and interfaces in graphql API handler for better readability
